Depend on the SAT module and rename to DSHARP
The new SAT module takes the repetitive interface out of this module.
The 01-dispatch.t test can hence be removed (it is contained in SAT).

The newly added `dsharp` routine conflicted with the exported class
SAT::Counter::dsharp, so I looked up how the project is actually called.
It's DSHARP -- not pretty but that's how we are going to call it too.

In general, the class name should always start with a capital letter,
while the routine form is all lowercase, like the programs usually are.